6.1 Создал PointController и перенес в него код, который отвечает за отрисовку точки маршрута, ее замену на форму редактирования и наоборот
