+++ src/controllers/trip.js
@@ -1,14 +1,12 @@
 import TripComponent from "../components/trip.js";
 import TripDayComponent from "../components/trip-day.js";
 import TripEmptyDayComponent from "../components/trip-day-empty.js";
-import TripPointComponent from "../components/trip-point.js";
-import TripPointEditComponent from "../components/trip-point-edit.js";
+import TripPointController from "./trip-point.js";
 import NoPointsComponent from "../components/no-points.js";
 import TripSortComponent, {SortType} from "../components/trip-sort.js";
 import dateFormat from "../utils/date-format.js";
 import constants from "../data/constants.js";
-import {render, replace} from "../utils/render.js";
-import {isEscapeEvent} from "../utils/events.js";
+import {render} from "../utils/render.js";
 
 const POINTS_CONTAINER_SELECTOR = `.trip-events__list`;
 
@@ -42,45 +40,11 @@ const groupPointsByStartDate = (orderedPoints) => {
   }, new Map());
 };
 
-const renderTripPoint = (container, point) => {
-  const editComponent = new TripPointEditComponent(point);
-  const viewComponent = new TripPointComponent(point);
-
-  const onEditButtonClick = () => {
-    replace(container, editComponent, viewComponent);
-  };
-
-  const hideEditForm = () => {
-    replace(container, viewComponent, editComponent);
-  };
-
-  const onCancelButtonClick = () => {
-    hideEditForm();
-  };
-
-  const onEditFormSubmit = (evt) => {
-    evt.preventDefault();
-    hideEditForm();
-  };
-
-  const onEscapeKeydown = (evt) => {
-    isEscapeEvent(evt, hideEditForm);
-  };
-
-  viewComponent.addOnEditButtonClickEvent(onEditButtonClick);
-
-  editComponent.addOnCancelButtonClickEvent(onCancelButtonClick);
-  editComponent.addOnFormSubmitEvent(onEditFormSubmit);
-  document.addEventListener(`keydown`, onEscapeKeydown);
-
-  render(container, viewComponent, constants.RENDER_POSITIONS.BEFORE_END);
-};
-
 const renderTripDay = (container, tripDayComponent, orderedPoints) => {
   const pointsContainer = tripDayComponent.getElement().querySelector(POINTS_CONTAINER_SELECTOR);
 
   for (let point of orderedPoints) {
-    renderTripPoint(pointsContainer, point);
+    new TripPointController(pointsContainer).render(point);
   }
 
   render(container, tripDayComponent, constants.RENDER_POSITIONS.BEFORE_END);
